log-producer-Python-bind9/requirements.txt
watchdog
kafka-python
cryptography
pyjks
pyinotify